Deprecating AdminController class in favor of EasyAdminController
