+++ tests/Fixtures/AppTestBundle/Controller/CustomTemplateParametersController.php
@@ -2,7 +2,7 @@
 
 namespace EasyCorp\Bundle\EasyAdminBundle\Tests\Fixtures\AppTestBundle\Controller;
 
-use EasyCorp\Bundle\EasyAdminBundle\Controller\AdminController as EasyAdminController;
+use EasyCorp\Bundle\EasyAdminBundle\Controller\EasyAdminController;
 
 class CustomTemplateParametersController extends EasyAdminController
 {
